EulerHS.CachedSqlDBQuery

Documentation

create :: forall (be :: Type) (beM :: Type -> Type) (table :: (Type -> Type) -> Type) (m :: Type -> Type). (HasCallStack, BeamRuntime be beM, BeamRunner beM, HasQBuilder be, Model be table, ToJSON (table Identity), Show (table Identity), MonadFlow m) => DBConfig beM -> table Identity -> Maybe Text -> m (Either DBError (table Identity))

createSql :: forall m be beM table. (HasCallStack, BeamRuntime be beM, BeamRunner beM, HasQBuilder be, Model be table, Show (table Identity), MonadFlow m) => DBConfig beM -> table Identity -> m (Either DBError (table Identity))

updateOne :: (HasCallStack, BeamRuntime be beM, BeamRunner beM, Model be table, HasQBuilder be, ToJSON (table Identity), Show (table Identity), MonadFlow m) => DBConfig beM -> Maybe Text -> [Set be table] -> Where be table -> m (Either DBError (table Identity))

Update an element matching the query to the new value. Cache the value at the given key if the DB update succeeds.

updateOneWoReturning :: (HasCallStack, BeamRuntime be beM, BeamRunner beM, Model be table, HasQBuilder be, MonadFlow m) => DBConfig beM -> Maybe Text -> [Set be table] -> Where be table -> m (Either DBError ())

updateOneSql :: forall m be beM table. (HasCallStack, BeamRuntime be beM, BeamRunner beM, Model be table, HasQBuilder be, Show (table Identity), MonadFlow m) => DBConfig beM -> [Set be table] -> Where be table -> m (DBResult (table Identity))

updateOneSqlWoReturning :: forall m be beM table. (HasCallStack, BeamRuntime be beM, BeamRunner beM, Model be table, HasQBuilder be, MonadFlow m) => DBConfig beM -> [Set be table] -> Where be table -> m (DBResult ())

updateExtended :: (HasCallStack, MonadFlow m, BeamRunner beM, BeamRuntime be beM) => DBConfig beM -> Maybe Text -> SqlUpdate be table -> m (Either DBError ())

Perform an arbitrary SqlUpdate. This will cache if successful.

findOne :: (HasCallStack, BeamRuntime be beM, BeamRunner beM, Model be table, HasQBuilder be, ToJSON (table Identity), FromJSON (table Identity), MonadFlow m) => DBConfig beM -> Maybe Text -> Where be table -> m (Either DBError (Maybe (table Identity)))

Find an element matching the query. Only uses the DB if the cache is empty. Caches the result using the given key.

findOneSql :: (HasCallStack, BeamRuntime be beM, BeamRunner beM, Model be table, HasQBuilder be, MonadFlow m) => DBConfig beM -> Where be table -> m (Either DBError (Maybe (table Identity)))

findAll :: (HasCallStack, BeamRuntime be beM, BeamRunner beM, Model be table, HasQBuilder be, ToJSON (table Identity), FromJSON (table Identity), MonadFlow m) => DBConfig beM -> Maybe Text -> Where be table -> m (Either DBError [table Identity])

Find all elements matching the query. Only uses the DB if the cache is empty. Caches the result using the given key. NOTE: Can't use the same key as findOne, updateOne or create since it's result is a list.

findAllSql :: (HasCallStack, BeamRuntime be beM, BeamRunner beM, Model be table, HasQBuilder be, MonadFlow m) => DBConfig beM -> Where be table -> m (Either DBError [table Identity])

findAllExtended :: forall beM be table m. (HasCallStack, MonadFlow m, FromBackendRow be (table Identity), BeamRunner beM, BeamRuntime be beM, FromJSON (table Identity), ToJSON (table Identity)) => DBConfig beM -> Maybe Text -> SqlSelect be (table Identity) -> m (Either DBError [table Identity])

Like findAll, but takes an explicit SqlSelect.

class SqlReturning (beM :: Type -> Type) (be :: Type) where

Create a new database entry with the given value. Cache the value if the DB insert succeeds.

Methods

createReturning :: forall (table :: (Type -> Type) -> Type) (m :: Type -> Type). (HasCallStack, BeamRuntime be beM, BeamRunner beM, HasQBuilder be, Model be table, ToJSON (table Identity), FromJSON (table Identity), Show (table Identity), MonadFlow m) => DBConfig beM -> table Identity -> Maybe Text -> m (Either DBError (table Identity))