EulerHS.Extra.AltValidation

Extra Validation

class Transform a b where

Represents Transformer from one type to another.

Methods

transform :: a -> Validation Errors b

mkValidator :: Text -> (t -> Bool) -> Validator t

Takes error message and predicate and returns validation function using default VErrorPayload

mkCustomValidator :: VErrorPayload -> Text -> (t -> Bool) -> Validator t

Takes error message and predicate and returns validation function using custom error

mkTransformer :: VErrorPayload -> (a -> Maybe b) -> Transformer a b

withCustomError :: VErrorPayload -> Validator a -> Validator a

Make a validator using a particular error message, original errors are ignored

type Transformer a b = a -> ReaderT Ctx (Either Errors) b

type Validator a = Transformer a a

Represents the value parameter validator.

type V a = Validation [VErrorPayload] a

type Errors = [VErrorPayload]

data VErrorPayload

Constructors

VErrorPayload

Fields

status :: Text

status_id :: Maybe Int

error_code :: Maybe Text

error_message :: Maybe Text

error_field :: Maybe Text

module Data.Validation

withField' :: forall (f :: Symbol) v r a. (HasField f r v, KnownSymbol f) => r -> Transformer v a -> Validation Errors a

Extract value and run validators on it This function is not working with HasField New withField without lens

withField :: forall (f :: Symbol) v r a. (HasField' f r v, KnownSymbol f) => r -> Transformer v a -> Validation Errors a

Extract value and run validators on it Old withField with lens

runParser :: forall a. ReaderT Ctx (Either Errors) a -> Text -> Validation Errors a

Run a custom parser

extractJust :: Transformer (Maybe a) a

Trying to extract the argument from Maybe type if value is Nothing then raise Failure

extractMaybeWithDefault :: a -> Transformer (Maybe a) a

guarded :: Text -> Bool -> ReaderT Ctx (Either Errors) ()

Guards computations by a validation rule

guardedCustom :: VErrorPayload -> Bool -> ReaderT Ctx (Either Errors) ()

Guards computations by a validation rule with a custom error

decode :: forall t. Read t => Transformer Text t

Trying to decode Text into a target type

decodeCustom :: forall t. Read t => VErrorPayload -> Transformer Text t

Trying to decode Text into a target type, use custom error

insideJust :: Transformer a b -> Transformer (Maybe a) (Maybe b)

parValidate :: [Validator a] -> Validator a