EulerHS.Extra.Secret

Description

Protecting sensitive data from being accidentally leaked by packing them into a Secret newtype wrapper.

Introduction

This module introduces Secret newtype wrapper. The idea is to put all sensitive fields in a Secret wrapper, so instead just Text / Int / etc fields we are going to work with Secret Text and so on.

data X = X {name :: Secret String, b :: Y}
  deriving stock (Generic)
  deriving anyclass (FromJSON)

  deriving instance Given SecretContext => ToJSON X
  deriving instance Given SecretContext => Show X
  deriving instance Given SecretContext => Store X

data Y = Y {c :: Int, money :: Secret Int}
  deriving stock (Generic)
  deriving anyclass (FromJSON, FromForm)

  deriving instance Given SecretContext => ToJSON Y
  deriving instance Given SecretContext => Show Y
  deriving instance Given SecretContext => Store Y

The Secret wrapper works as follows:

you can put a value into a Secret using makeSecret function

in order to extract a value you have to provide 'Given SecretContext` instance

Let's start with creating:

exampleX :: X
exampleX = X { name = makeSecret "a", b = exampleY }

exampleY :: Y
exampleY = Y {c = 10, money = makeSecret 20}

Potentially dangerous instances including Show, ToJSON and Store also require 'Given SecretContext'. In order not to pass the context explicitly this module leverages Data.Reflection machinery to pass around the context. You don't need to understand how reifyreflect and givegiven is implemented but you have to understand this:

>>> given (42 :: Int) $ given @Int
42

Let's take a look at more examples:

>>> print $ give SafelyHideSecrets (A.encode exampleX)
"{\"name\":\"***secret***\",\"b\":{\"money\":\"***secret***\",\"c\":10}}"
>>> print $ give RiskyShowSecrets (A.encode exampleX)
"{\"name\":\"a\",\"b\":{\"money\":20,\"c\":10}}"
>>> print $ give SafelyHideSecrets (show exampleX :: String)
"X {name = ***secret***, b = Y {c = 10, money = ***secret***}}"
>>> print $ give RiskyShowSecrets (show exampleX :: String)
"X {name = \"a\", b = Y {c = 10, money = 20}}"

TODO: Is it needed? FromJSON, FromForm, FromHttpApiData and Stores parsing fails if encounters secreted data:

>>> let secretJson = A.decode @X $ give SafelyHideSecrets $ A.encode exampleX
"*** Exception: FromJSON (Secret a): trying to decode secret data

For changing Secret data just use Secret as a monad without extraction:

f :: Secret Int -> Secret String
f secret = do
 x <- secret
 pure $ show $ x + 1

Core part

data SecretContext

Secret context to annotate risky/safely usage of Secret type

Constructors

RiskyShowSecrets

SafelyHideSecrets

data Secret (a :: Type)

Secret data type is used to hide any sensitive data stored in fields in API and domain types.

We refused use the Traversable instance intentionally since it opens backdoors like 'traverse print (Secret password)'. The same is the case with Foldable leading to extraction data, e.g.:

sneaky :: Show a => Secret a -> String
sneaky = foldMap show

Construction and extraction

makeSecret :: a -> Secret a

Just makes a secret.

type ExtractSecrets = ExtractSecrets' ()

A constraint to track functions which requires access to sensitive data.

extractSecret :: ExtractSecrets => Secret a -> a

A function to extract a value in ExtractSecrets context.

elimExtractSecrets :: forall a. (ExtractSecrets => a) -> a

Eliminates ExtractSecrets constraint.

>>> elimExtractSecrets $ extractSecret  $ Secret ("foo" :: String)
"foo"

toBuilder :: Given SecretContext => Secret a -> (a -> Builder) -> Builder

TODO do we need it? If yes, why not to have extract :: Given SecretContext => Secret a -> Maybe a?

unsafeExtractSecret :: Secret a -> a

Deprecated: use extractSecret instead; for backward compatibility only

The last resort to extract sensitive data. Avoid to use it as as long as possible! It is NOT for show or logs!

Please, annotate all uses of unsafeExtractSecre with a comment that says on why this was done.

sequence-like transformations

sequenceSecretMaybe :: Secret (Maybe a) -> Maybe (Secret a)

sequenceMaybeSecret :: Maybe (Secret a) -> Secret (Maybe a)

sequenceSecretEither :: (Given SecretContext, IsString e) => Secret (Either e a) -> Either e (Secret a)

sequenceEitherSecret :: Either e (Secret a) -> Secret (Either e a)

Servant stuff

The Secret is designed to be used with servant, one can use it in request and response types. Nothing is needed to protect requests since packing into Secret is a straightforward operation. For responses just use ProtectedJSON instead of stock JSON in your API-type and use via deriving for ToJSON instance using DiscloseSecret newtype:

type Api =
  "path"
  :> ReqBody '[JSON] Req        -- no changes here
  :> Post '[ProtectedJSON] Foo  -- uses ProtectedJSON

data Foo = Foo { bar :: Secret Text }

deriving stock instance Eq InitTxnResp,
deriving stock instance Generic InitTxnResp
deriving anyclass instance Given SecretContext => ToJSON Foo
deriving via (DiscloseSecret Foo) instance MimeRender ProtectedJSON Foo

data ProtectedJSON

A promoted type constructor to use instead of the regular JSON kind

newtype DiscloseSecret a

A newtype wrapper to use in via deriving

Constructors

DiscloseSecret a

Re-exports

Some of Data.Reflection are get re-exported.

class Given a

This is a version of Reifies that allows for only a single value.

This is easier to work with than Reifies and permits extended defaulting, but it only offers a single reflected value of a given type at a time.

Minimal complete definition

given

give :: a -> (Given a => r) -> r

Reify a value into an instance to be recovered with given.

You should only give a single value for each type. If multiple instances are in scope, then the behavior is implementation defined.