EulerHS.Extra.Validation

Extra Validation

class Transform a b where

Represents Transformer from one type to another.

Methods

transform :: a -> Validation Errors b

mkValidator :: Text -> (t -> Bool) -> Validator t

Takes error message and predicate and return validation function

mkTransformer :: Text -> (a -> Maybe b) -> Transformer a b

type Transformer a b = a -> ReaderT Ctx (Either Errors) b

type Validator a = Transformer a a

Represents the value parameter validator.

type V a = Validation [Text] a

type Errors = [Text]

module Data.Validation

withField' :: forall (f :: Symbol) v r a. (HasField f r v, KnownSymbol f) => r -> Transformer v a -> Validation Errors a

Extract value and run validators on it New One

withField :: forall (f :: Symbol) v r a. (HasField' f r v, KnownSymbol f) => r -> Transformer v a -> Validation Errors a

runParser :: forall a. ReaderT Ctx (Either Errors) a -> Text -> Validation Errors a

Run parser

extractJust :: Transformer (Maybe a) a

Trying to extract the argument from Maybe type if value is Nothing then raise Failure

extractMaybeWithDefault :: a -> Transformer (Maybe a) a

guarded :: Text -> Bool -> ReaderT Ctx (Either Errors) ()

decode :: forall t. Read t => Transformer Text t

Trying to decode Text to target type

insideJust :: Transformer a b -> Transformer (Maybe a) (Maybe b)

parValidate :: [Validator a] -> Validator a