EulerHS.Language

Documentation

keyToSlot :: ByteString -> HashSlot

type KVDB next = ExceptT KVDBReply (F KVDBF) next

data KVDBF next

Constructors

KV (KeyValueF KVDBAnswer next)

TX (TransactionF next)

data TransactionF next where

Constructors

MultiExec :: KVDBTx (Queued a) -> (KVDBAnswer (TxResult a) -> next) -> TransactionF next

MultiExecWithHash :: ByteString -> KVDBTx (Queued a) -> (KVDBAnswer (TxResult a) -> next) -> TransactionF next

type KVDBTx = F (KeyValueF Queued)

data KeyValueF f next where

Constructors

Set :: KVDBKey -> KVDBValue -> (f KVDBStatus -> next) -> KeyValueF f next

SetEx :: KVDBKey -> KVDBDuration -> KVDBValue -> (f KVDBStatus -> next) -> KeyValueF f next

SetOpts :: KVDBKey -> KVDBValue -> KVDBSetTTLOption -> KVDBSetConditionOption -> (f Bool -> next) -> KeyValueF f next

Get :: KVDBKey -> (f (Maybe ByteString) -> next) -> KeyValueF f next

Exists :: KVDBKey -> (f Bool -> next) -> KeyValueF f next

Del :: [KVDBKey] -> (f Integer -> next) -> KeyValueF f next

Expire :: KVDBKey -> KVDBDuration -> (f Bool -> next) -> KeyValueF f next

Incr :: KVDBKey -> (f Integer -> next) -> KeyValueF f next

HSet :: KVDBKey -> KVDBField -> KVDBValue -> (f Bool -> next) -> KeyValueF f next

HGet :: KVDBKey -> KVDBField -> (f (Maybe ByteString) -> next) -> KeyValueF f next

XAdd :: KVDBStream -> KVDBStreamEntryIDInput -> [KVDBStreamItem] -> (f KVDBStreamEntryID -> next) -> KeyValueF f next

XLen :: KVDBStream -> (f Integer -> next) -> KeyValueF f next

SAdd :: KVDBKey -> [KVDBValue] -> (f Integer -> next) -> KeyValueF f next

SMem :: KVDBKey -> KVDBKey -> (f Bool -> next) -> KeyValueF f next

Raw :: RedisResult a => [ByteString] -> (f a -> next) -> KeyValueF f next

type KVDBStreamItem = (ByteString, ByteString)

data KVDBStreamEntryIDInput

Constructors

EntryID KVDBStreamEntryID

AutoID

data KVDBStreamEntryID

Constructors

KVDBStreamEntryID Integer Integer

type KVDBStream = ByteString

type KVDBMessage = ByteString

type KVDBChannel = ByteString

type KVDBField = ByteString

type KVDBDuration = Integer

type KVDBValue = ByteString

type KVDBKey = ByteString

data KVDBSetConditionOption

Constructors

SetAlways

SetIfExist

SetIfNotExist

data KVDBSetTTLOption

Constructors

NoTTL

Seconds Integer

Milliseconds Integer

setTx :: KVDBKey -> KVDBValue -> KVDBTx (Queued KVDBStatus)

Set the value of a key. Transaction version.

setexTx :: KVDBKey -> KVDBDuration -> KVDBValue -> KVDBTx (Queued KVDBStatus)

Set the value and ttl of a key. Transaction version.

hsetTx :: KVDBKey -> KVDBField -> KVDBValue -> KVDBTx (Queued Bool)

Set the value of a hash field. Transaction version.

getTx :: KVDBKey -> KVDBTx (Queued (Maybe ByteString))

Get the value of a key. Transaction version.

hgetTx :: KVDBKey -> KVDBField -> KVDBTx (Queued (Maybe ByteString))

Get the value of a hash field. Transaction version.

delTx :: [KVDBKey] -> KVDBTx (Queued Integer)

Delete a keys. Transaction version.

expireTx :: KVDBKey -> KVDBDuration -> KVDBTx (Queued Bool)

Set a key's time to live in seconds. Transaction version.

xaddTx :: KVDBStream -> KVDBStreamEntryIDInput -> [KVDBStreamItem] -> KVDBTx (Queued KVDBStreamEntryID)

xlenTx :: KVDBStream -> KVDBTx (Queued Integer)

set :: KVDBKey -> KVDBValue -> KVDB KVDBStatus

Set the value of a key

setex :: KVDBKey -> KVDBDuration -> KVDBValue -> KVDB KVDBStatus

Set the value and ttl of a key.

setOpts :: KVDBKey -> KVDBValue -> KVDBSetTTLOption -> KVDBSetConditionOption -> KVDB Bool

get :: KVDBKey -> KVDB (Maybe ByteString)

Get the value of a key

exists :: KVDBKey -> KVDB Bool

Determine if a key exists

del :: [KVDBKey] -> KVDB Integer

Delete a keys

expire :: KVDBKey -> KVDBDuration -> KVDB Bool

Set a key's time to live in seconds

incr :: KVDBKey -> KVDB Integer

Increment the integer value of a key by one

hset :: KVDBKey -> KVDBField -> KVDBValue -> KVDB Bool

Set the value of a hash field

hget :: KVDBKey -> KVDBField -> KVDB (Maybe ByteString)

Get the value of a hash field

xadd :: KVDBStream -> KVDBStreamEntryIDInput -> [KVDBStreamItem] -> KVDB KVDBStreamEntryID

xlen :: KVDBStream -> KVDB Integer

sadd :: KVDBKey -> [KVDBValue] -> KVDB Integer

Add one or more members to a set

sismember :: KVDBKey -> KVDBKey -> KVDB Bool

multiExec :: KVDBTx (Queued a) -> KVDB (TxResult a)

Run commands inside a transaction(suited only for standalone redis setup).

multiExecWithHash :: ByteString -> KVDBTx (Queued a) -> KVDB (TxResult a)

Run commands inside a transaction(suited only for cluster redis setup).

rawRequest :: RedisResult a => [ByteString] -> KVDB a

Perform a raw call against the underlying Redis data store. This is definitely unsafe, and should only be used if you know what you're doing.

See also: The Hedis function this is based on.

Since: 2.0.3.2

type Logger = F LoggerMethod

data LoggerMethod next where

Language for logging.

Constructors

LogMessage :: LogLevel -> !Tag -> !Message -> (() -> next) -> LoggerMethod next
Log message with a predefined level.

logMessage' :: forall tag. (Typeable tag, Show tag) => LogLevel -> tag -> Message -> Logger ()

type PubSub = F PubSubF

data PubSubF next

Constructors

Publish Channel Payload (Either KVDBReply Integer -> next)

Subscribe [Channel] MessageCallback (IO () -> next)

PSubscribe [ChannelPattern] PMessageCallback (IO () -> next)

newtype Payload

Constructors

Payload ByteString

newtype ChannelPattern

Constructors

ChannelPattern ByteString

newtype Channel

Constructors

Channel ByteString

data SqlDBMethodF (beM :: Type -> Type) next where

Constructors

SqlDBMethod :: HasCallStack => (NativeSqlConn -> (Text -> IO ()) -> IO a) -> (a -> next) -> SqlDBMethodF beM next

SqlThrowException :: (HasCallStack, Exception e) => e -> (a -> next) -> SqlDBMethodF beM next

type SqlDB beM = F (SqlDBMethodF beM)

sqlThrowException :: forall a e beM. (HasCallStack, Exception e) => e -> SqlDB beM a

findRows :: (HasCallStack, BeamRunner beM, BeamRuntime be beM, FromBackendRow be a) => SqlSelect be a -> SqlDB beM [a]

Select many

findRow :: (HasCallStack, BeamRunner beM, BeamRuntime be beM, FromBackendRow be a) => SqlSelect be a -> SqlDB beM (Maybe a)

Select one

insertRows :: (HasCallStack, BeamRunner beM, BeamRuntime be beM) => SqlInsert be table -> SqlDB beM ()

Insert

insertRowsReturningList :: (HasCallStack, Beamable table, FromBackendRow be (table Identity), BeamRuntime be beM, BeamRunner beM) => SqlInsert be table -> SqlDB beM [table Identity]

Insert returning list

updateRows :: (HasCallStack, BeamRunner beM, BeamRuntime be beM) => SqlUpdate be table -> SqlDB beM ()

Update

updateRowsReturningList :: (HasCallStack, BeamRunner beM, BeamRuntime be beM, Beamable table, FromBackendRow be (table Identity)) => SqlUpdate be table -> SqlDB beM [table Identity]

Update returning list

deleteRows :: (HasCallStack, BeamRunner beM, BeamRuntime be beM) => SqlDelete be table -> SqlDB beM ()

Delete

deleteRowsReturningListPG :: (HasCallStack, Beamable table, FromBackendRow Postgres (table Identity)) => SqlDelete Postgres table -> SqlDB Pg [table Identity]

updateRowsReturningListPG :: (HasCallStack, Beamable table, FromBackendRow Postgres (table Identity)) => SqlUpdate Postgres table -> SqlDB Pg [table Identity]

insertRowReturningMySQL :: (HasCallStack, FromBackendRow MySQL (table Identity)) => SqlInsert MySQL table -> SqlDB MySQLM (Maybe (table Identity))

data AppException

An app-specific exception.

Since: 2.1.0.1

Constructors

SqlDBConnectionFailedException Text

KVDBConnectionFailedException Text

getCurrentTimeUTC :: MonadFlow m => m LocalTime

Retrieves the current UTC time, but as a LocalTime.

Since: 2.1.0.1

getCurrentDateInSeconds :: MonadFlow m => m Int

Retrieves the current POSIX time, rounded to seconds.

Since: 2.1.0.1

getCurrentDateInMillis :: MonadFlow m => m Int

Retrieves the current POSIX time, rounded to milliseconds.

Since: 2.1.0.1

getCurrentDateStringWithSecOffset :: MonadFlow m => Int -> m Text

Given a number of seconds as an offset, return a date string, in the format YYYY-MM-ddTHH:MM:SSZ, representing the current time, offset by the specified number of seconds.

Since: 2.1.0.1

throwOnFailedWithLog :: (HasCallStack, Show e, MonadFlow m) => Either e a -> (Text -> AppException) -> Text -> m ()

Transforms a Left result into an exception, logging this outcome. Does nothing on a Right.

Since: 2.1.0.1

checkFailedWithLog :: (HasCallStack, Show e, MonadFlow m) => Either e a -> Text -> m ()

logInfoT :: forall (m :: Type -> Type). (HasCallStack, MonadFlow m) => Text -> Text -> m ()

As logInfo, but specialized for logging Text tags.

Since: 2.1.0.1

logErrorT :: forall (m :: Type -> Type). (HasCallStack, MonadFlow m) => Text -> Text -> m ()

As logError, but specialized for logging Text tags.

Since: 2.1.0.1

logDebugT :: forall (m :: Type -> Type). (HasCallStack, MonadFlow m) => Text -> Text -> m ()

As logDebug, but specialized for logging Text tags.

Since: 2.1.0.1

logWarningT :: forall (m :: Type -> Type). (HasCallStack, MonadFlow m) => Text -> Text -> m ()

As logWarning, but specialized for logging Text tags.

Since: 2.1.0.1

withDB :: (HasCallStack, MonadFlow m, BeamRunner beM, BeamRuntime be beM) => DBConfig beM -> SqlDB beM a -> m a

Creates a connection and runs a DB operation. Throws on connection failure or if the operation fails; this will log if either of these things happens.

NOTE: This does not run inside a transaction.

Since: 2.1.0.1

withDBTransaction :: (HasCallStack, MonadFlow m, BeamRunner beM, BeamRuntime be beM) => DBConfig beM -> SqlDB beM a -> m a

As withDB, but runs inside a transaction.

Since: 2.1.0.1

insertRow :: (HasCallStack, MonadFlow m, BeamRunner beM, BeamRuntime be beM, Beamable table, FromBackendRow be (table Identity)) => DBConfig beM -> SqlInsert be table -> m (Either Text (table Identity))

Inserts several rows, returning the first successful inserted result. Use this function with care: if your insert ends up inserting nothing successfully, this will return a Left.

Since: 2.1.0.1

unsafeInsertRow :: (HasCallStack, MonadFlow m, BeamRunner beM, BeamRuntime be beM, Beamable table, FromBackendRow be (table Identity), Exception e) => e -> DBConfig beM -> SqlInsert be table -> m (table Identity)

As insertRow, but instead throws the provided exception on failure. Will also log in such a case.

Since: 2.1.0.1

insertRowMySQL :: (HasCallStack, MonadFlow m, FromBackendRow MySQL (table Identity)) => DBConfig MySQLM -> SqlInsert MySQL table -> m (Either Text (table Identity))

MySQL-specific version of insertRow.

Since: 2.1.0.1

unsafeInsertRowMySQL :: (HasCallStack, MonadFlow m, FromBackendRow MySQL (table Identity), Exception e) => e -> DBConfig MySQLM -> SqlInsert MySQL table -> m (table Identity)

MySQL-specific version of unsafeInsertRow.

Since: 2.1.0.1

getOrInitSqlConn :: (HasCallStack, MonadFlow m) => DBConfig beM -> m (DBResult (SqlConn beM))

Get existing SQL connection, or init a new connection.

getOrInitKVDBConn :: (HasCallStack, MonadFlow m) => KVDBConfig -> m (KVDBAnswer KVDBConn)

Get existing Redis connection, or init a new connection.

rExpire :: (HasCallStack, Integral t, MonadFlow m) => RedisName -> TextKey -> t -> m (Either KVDBReply Bool)

rExpireB :: (HasCallStack, Integral t, MonadFlow m) => RedisName -> ByteKey -> t -> m (Either KVDBReply Bool)

rDel :: (HasCallStack, MonadFlow m) => RedisName -> [TextKey] -> m (Either KVDBReply Integer)

rDelB :: (HasCallStack, MonadFlow m) => RedisName -> [ByteKey] -> m (Either KVDBReply Integer)

rExists :: (HasCallStack, MonadFlow m) => RedisName -> TextKey -> m (Either KVDBReply Bool)

rExistsB :: (HasCallStack, MonadFlow m) => RedisName -> ByteKey -> m (Either KVDBReply Bool)

rExistsT :: (HasCallStack, MonadFlow m) => RedisName -> TextKey -> m (Either KVDBReply Bool)

rHget :: (HasCallStack, FromJSON v, MonadFlow m) => RedisName -> TextKey -> TextField -> m (Maybe v)

rHgetB :: (HasCallStack, MonadFlow m) => Text -> ByteKey -> ByteField -> m (Maybe ByteValue)

rHset :: (HasCallStack, ToJSON v, MonadFlow m) => RedisName -> TextKey -> TextField -> v -> m (Either KVDBReply Bool)

rHsetB :: (HasCallStack, MonadFlow m) => RedisName -> ByteKey -> ByteField -> ByteValue -> m (Either KVDBReply Bool)

rIncr :: (HasCallStack, MonadFlow m) => RedisName -> TextKey -> m (Either KVDBReply Integer)

rIncrB :: (HasCallStack, MonadFlow m) => RedisName -> ByteKey -> m (Either KVDBReply Integer)

rSet :: (HasCallStack, ToJSON v, MonadFlow m) => RedisName -> TextKey -> v -> m (Either KVDBReply KVDBStatus)

rSetB :: (HasCallStack, MonadFlow m) => Text -> ByteKey -> ByteValue -> m (Either KVDBReply KVDBStatus)

rSetT :: (HasCallStack, MonadFlow m) => RedisName -> TextKey -> Text -> m (Either KVDBReply KVDBStatus)

rGetB :: (HasCallStack, MonadFlow m) => RedisName -> ByteKey -> m (Maybe ByteValue)

rGet :: (HasCallStack, FromJSON v, MonadFlow m) => RedisName -> TextKey -> m (Maybe v)

rGetT :: (HasCallStack, MonadFlow m) => Text -> Text -> m (Maybe Text)

rSetex :: (HasCallStack, ToJSON v, Integral t, MonadFlow m) => RedisName -> TextKey -> v -> t -> m (Either KVDBReply KVDBStatus)

rSetexB :: (HasCallStack, Integral t, MonadFlow m) => RedisName -> ByteKey -> ByteValue -> t -> m (Either KVDBReply KVDBStatus)

rSetexT :: (HasCallStack, ToJSON v, Integral t, MonadFlow m) => RedisName -> TextKey -> v -> t -> m (Either KVDBReply KVDBStatus)

rSetexBulk :: (HasCallStack, ToJSON v, Integral t, MonadFlow m) => RedisName -> Map TextKey v -> t -> m (Either KVDBReply ())

rSetexBulkB :: (HasCallStack, Integral t, MonadFlow m) => RedisName -> Map ByteKey ByteValue -> t -> m (Either KVDBReply ())

rSetOpts :: (HasCallStack, ToJSON v, MonadFlow m) => RedisName -> TextKey -> v -> KVDBSetTTLOption -> KVDBSetConditionOption -> m (Either KVDBReply Bool)

rSetOptsB :: (HasCallStack, MonadFlow m) => RedisName -> ByteKey -> ByteValue -> KVDBSetTTLOption -> KVDBSetConditionOption -> m (Either KVDBReply Bool)

rSetOptsT :: (HasCallStack, MonadFlow m) => RedisName -> TextKey -> Text -> KVDBSetTTLOption -> KVDBSetConditionOption -> m (Either KVDBReply Bool)

rSadd :: (HasCallStack, MonadFlow m) => RedisName -> KVDBKey -> [KVDBValue] -> m (Either KVDBReply Integer)

rSismember :: (HasCallStack, MonadFlow m) => RedisName -> KVDBKey -> KVDBValue -> m (Either KVDBReply Bool)

withLoggerContext :: (HasCallStack, MonadFlow m) => (LogContext -> LogContext) -> Flow a -> m a

updateLoggerContext :: (LogContext -> LogContext) -> FlowRuntime -> FlowRuntime

data Flow (a :: Type)

data FlowMethod (next :: Type) where

Flow language.

Constructors

LookupHTTPManager :: HasCallStack => Maybe ManagerSelector -> (Maybe Manager -> next) -> FlowMethod next

GetHTTPManager :: HasCallStack => HTTPClientSettings -> (Manager -> next) -> FlowMethod next

CallServantAPI :: HasCallStack => Manager -> BaseUrl -> EulerClient a -> (Either ClientError a -> next) -> FlowMethod next

CallHTTP :: HasCallStack => HTTPRequest -> Manager -> (Either Text HTTPResponse -> next) -> FlowMethod next

EvalLogger :: HasCallStack => Logger a -> (a -> next) -> FlowMethod next

RunIO :: HasCallStack => Text -> IO a -> (a -> next) -> FlowMethod next

WithRunFlow :: HasCallStack => ((forall x. Flow x -> IO x) -> IO next) -> FlowMethod next

GetOption :: HasCallStack => Text -> (Maybe a -> next) -> FlowMethod next

SetOption :: HasCallStack => Text -> a -> (() -> next) -> FlowMethod next

DelOption :: HasCallStack => Text -> (() -> next) -> FlowMethod next

GenerateGUID :: HasCallStack => (Text -> next) -> FlowMethod next

RunSysCmd :: HasCallStack => String -> (String -> next) -> FlowMethod next

Fork :: HasCallStack => Description -> ForkGUID -> Flow a -> (Awaitable (Either Text a) -> next) -> FlowMethod next

Await :: HasCallStack => Maybe Microseconds -> Awaitable (Either Text a) -> (Either AwaitingError a -> next) -> FlowMethod next

ThrowException :: forall a e next. (HasCallStack, Exception e) => e -> (a -> next) -> FlowMethod next

CatchException :: forall a e next. (HasCallStack, Exception e) => Flow a -> (e -> Flow a) -> (a -> next) -> FlowMethod next

Mask :: forall b next. HasCallStack => ((forall a. Flow a -> Flow a) -> Flow b) -> (b -> next) -> FlowMethod next

UninterruptibleMask :: forall b next. HasCallStack => ((forall a. Flow a -> Flow a) -> Flow b) -> (b -> next) -> FlowMethod next

GeneralBracket :: forall a b c next. HasCallStack => Flow a -> (a -> ExitCase b -> Flow c) -> (a -> Flow b) -> ((b, c) -> next) -> FlowMethod next

RunSafeFlow :: HasCallStack => SafeFlowGUID -> Flow a -> (Either Text a -> next) -> FlowMethod next

InitSqlDBConnection :: HasCallStack => DBConfig beM -> (DBResult (SqlConn beM) -> next) -> FlowMethod next

DeInitSqlDBConnection :: HasCallStack => SqlConn beM -> (() -> next) -> FlowMethod next

GetSqlDBConnection :: HasCallStack => DBConfig beM -> (DBResult (SqlConn beM) -> next) -> FlowMethod next

InitKVDBConnection :: HasCallStack => KVDBConfig -> (KVDBAnswer KVDBConn -> next) -> FlowMethod next

DeInitKVDBConnection :: HasCallStack => KVDBConn -> (() -> next) -> FlowMethod next

GetKVDBConnection :: HasCallStack => KVDBConfig -> (KVDBAnswer KVDBConn -> next) -> FlowMethod next

RunDB :: HasCallStack => SqlConn beM -> SqlDB beM a -> Bool -> (DBResult a -> next) -> FlowMethod next

RunKVDB :: HasCallStack => Text -> KVDB a -> (KVDBAnswer a -> next) -> FlowMethod next

RunPubSub :: HasCallStack => PubSub a -> (a -> next) -> FlowMethod next

WithModifiedRuntime :: HasCallStack => (FlowRuntime -> FlowRuntime) -> Flow a -> (a -> next) -> FlowMethod next

class MonadMask m => MonadFlow m where

MonadFlow implementation for the Flow Monad. This allows implementation of MonadFlow for ReaderT and other monad transformers.

Omit forkFlow as this will break some monads like StateT (you can lift this manually if you know what you're doing)

Minimal complete definition

callServantAPI, callAPIUsingManager, lookupHTTPManager, getHTTPManager, callHTTPUsingManager, evalLogger', runIO', getOption, setOption, delOption, generateGUID, runSysCmd, initSqlDBConnection, deinitSqlDBConnection, getSqlDBConnection, initKVDBConnection, deinitKVDBConnection, getKVDBConnection, runDB, runTransaction, await, runSafeFlow, runKVDB, runPubSub, publish, subscribe, psubscribe, withModifiedRuntime

Methods

callServantAPI

Arguments

:: HasCallStack

=> Maybe ManagerSelector
name of the connection manager to be used

-> BaseUrl
remote url BaseUrl

-> EulerClient a
servant client EulerClient

-> m (Either ClientError a)
result

Method for calling external HTTP APIs using the facilities of servant-client. Allows to specify what manager should be used. If no manager found, HttpManagerNotFound will be returne (as part of ConnectionError).

Thread safe, exception free.

Takes remote url, servant client for this endpoint and returns either client error or result.

data User = User { firstName :: String, lastName :: String , userGUID :: String}
  deriving (Generic, Show, Eq, ToJSON, FromJSON )

data Book = Book { author :: String, name :: String }
  deriving (Generic, Show, Eq, ToJSON, FromJSON )

type API = "user" :> Get '[JSON] User
      :<|> "book" :> Get '[JSON] Book

api :: HasCallStack => Proxy API
api = Proxy

getUser :: HasCallStack => EulerClient User
getBook :: HasCallStack => EulerClient Book
(getUser :<|> getBook) = client api

url = BaseUrl Http "localhost" port ""

myFlow = do
  book <- callServantAPI url getBook
  user <- callServantAPI url getUser

callAPIUsingManager :: HasCallStack => Manager -> BaseUrl -> EulerClient a -> m (Either ClientError a)

lookupHTTPManager :: (HasCallStack, MonadFlow m) => Maybe ManagerSelector -> m (Maybe Manager)

getHTTPManager :: HasCallStack => HTTPClientSettings -> m Manager

callHTTPUsingManager :: HasCallStack => Manager -> HTTPRequest -> m (Either Text HTTPResponse)

Method for calling external HTTP APIs without bothering with types.

Thread safe, exception free.

evalLogger' :: HasCallStack => Logger a -> m a

Evaluates a logging action.

runIO' :: HasCallStack => Text -> IO a -> m a

The same as runIO, but accepts a description which will be written into the ART recordings for better clarity.

Warning. This method is dangerous and should be used wisely.

myFlow = do
  content <- runIO' "reading from file" $ readFromFile file
  logDebugT "content id" $ extractContentId content
  pure content

getOption :: forall k v. (HasCallStack, OptionEntity k v) => k -> m (Maybe v)

Gets stored a typed option by a typed key.

Thread safe, exception free.

setOption :: forall k v. (HasCallStack, OptionEntity k v) => k -> v -> m ()

delOption :: forall k v. (HasCallStack, OptionEntity k v) => k -> m ()

Deletes a typed option using a typed key.

generateGUID :: HasCallStack => m Text

Generate a version 4 UUIDs as specified in RFC 4122 e.g. 25A8FC2A-98F2-4B86-98F6-84324AF28611.

Thread safe, exception free.

runSysCmd :: HasCallStack => String -> m String

Runs system command and returns its output.

Warning. This method is dangerous and should be used wisely.

myFlow = do
  currentDir <- runSysCmd "pwd"
  logInfoT "currentDir" $ toText currentDir
  ...

initSqlDBConnection :: HasCallStack => DBConfig beM -> m (DBResult (SqlConn beM))

Inits an SQL connection using a config.

Returns an error (Left $ T.DBError T.ConnectionAlreadyExists msg) if the connection already exists for this config.

Thread safe, exception free.

deinitSqlDBConnection :: HasCallStack => SqlConn beM -> m ()

Deinits an SQL connection. Does nothing if the connection is not found (might have been closed earlier).

Thread safe, exception free.

getSqlDBConnection :: HasCallStack => DBConfig beM -> m (DBResult (SqlConn beM))

Gets the existing connection.

Returns an error (Left $ T.DBError T.ConnectionDoesNotExist) if the connection does not exist.

Thread safe, exception free.

initKVDBConnection :: HasCallStack => KVDBConfig -> m (KVDBAnswer KVDBConn)

Inits a KV DB connection using a config.

Returns an error (Left $ KVDBError KVDBConnectionAlreadyExists msg) if the connection already exists.

Thread safe, exception free.

deinitKVDBConnection :: HasCallStack => KVDBConn -> m ()

Deinits the given KV DB connection. Does nothing if the connection is not found (might have been closed earlier).

Thread safe, exception free.

getKVDBConnection :: HasCallStack => KVDBConfig -> m (KVDBAnswer KVDBConn)

Get the existing connection.

runDB :: (HasCallStack, BeamRunner beM, BeamRuntime be beM) => SqlConn beM -> SqlDB beM a -> m (DBResult a)

Evaluates SQL DB operations outside of any transaction. It's possible to have a chain of SQL DB calls (within the SqlDB language). These chains will be executed as a single transaction.

Thread safe, exception free.

The underlying library is beam which allows to access 3 different SQL backends. See TUTORIAL.md, README.md and QueryExamplesSpec.hs for more info.

myFlow :: HasCallStack => L.Flow (T.DBResult (Maybe User))
myFlow = do
  connection <- L.initSqlDBConnection postgresCfg

  res <- L.runDB connection $ do
    let predicate1 User {..} = _userFirstName ==. B.val_ "John"

    L.updateRows $ B.update (_users eulerDb)
      (\User {..} -> mconcat
        [ _userFirstName <-. B.val_ "Leo"
        , _userLastName  <-. B.val_ "San"
        ]
      )
      predicate1

    let predicate2 User {..} = _userFirstName ==. B.val_ "Leo"
    L.findRow
      $ B.select
      $ B.limit_ 1
      $ B.filter_ predicate2
      $ B.all_ (_users eulerDb)

  L.deinitSqlDBConnection connection
  pure res

runTransaction :: (HasCallStack, BeamRunner beM, BeamRuntime be beM) => SqlConn beM -> SqlDB beM a -> m (DBResult a)

Like runDB but runs inside a SQL transaction.

await :: HasCallStack => Maybe Microseconds -> Awaitable (Either Text a) -> m (Either AwaitingError a)

Await for some a result from the flow. If the timeout is Nothing than the operation is blocking. If the timeout is set then the internal mechanism tries to do several (10) checks for the result. Can return earlier if the result became available. Returns either an Awaiting error or a result.

Warning. There are no guarantees of a proper thread delaying here.

Thread safe, exception free.

| mbMcs == Nothing: infinite awaiting. | mbMcs == Just (Microseconds n): await for approximately n seconds. Awaiting may succeed ealier.

myFlow1 = do
  logInfoT "myflow1" "logFromMyFlow1"
  pure 10

myFlow2 = do
  awaitable <- forkFlow' "myFlow1 fork" myFlow1
  await Nothing awaitable

throwException :: forall a e. (HasCallStack, Exception e) => e -> m a

Throw a given exception.

It's possible to catch this exception using runSafeFlow method.

Thread safe. Exception throwing.

myFlow = do
  res <- authAction
  case res of
    Failure reason -> throwException err403 {errBody = reason}
    Success -> ...

throwExceptionWithoutCallStack :: forall a e. (HasCallStack, Exception e) => e -> m a

runSafeFlow :: HasCallStack => Flow a -> m (Either Text a)

Run a flow safely with catching all the exceptions from it. Returns either a result or the exception turned into a text message.

This includes ususal instances of the Exception type class, error exception and custom user exceptions thrown by the throwException method.

Thread safe, exception free.

myFlow = runSafeFlow $ throwException err403 {errBody = reason}

myFlow = do
  eitherContent <- runSafeFlow $ runIO $ readFromFile file
  case eitherContent of
    Left err -> ...
    Right content -> ...

runKVDB

Arguments

:: HasCallStack

=> Text

-> KVDB a
KVDB action

-> m (KVDBAnswer a)

Execute kvdb actions.

Thread safe, exception free.

myFlow = do
  kvres <- L.runKVDB $ do
    set "aaa" "bbb"
    res <- get "aaa"
    del ["aaa"]
    pure res

runPubSub :: HasCallStack => PubSub a -> m a

publish

Arguments

:: HasCallStack

=> Channel
Channel in which payload will be send

-> Payload
Payload

-> m (Either KVDBReply Integer)
Number of subscribers received payload

Publish payload to channel.

subscribe

Arguments

:: HasCallStack

=> [Channel]
List of channels to subscribe

-> MessageCallback
Callback function.

-> m (Flow ())
Inner flow is a canceller of current subscription

Subscribe to all channels from list. Note: Subscription won't be unsubscribed automatically on thread end. Use canceller explicitly to cancel subscription

psubscribe

Arguments

:: HasCallStack

=> [ChannelPattern]
List of channels to subscribe (wit respect to patterns supported by redis)

-> PMessageCallback
Callback function

-> m (Flow ())
Inner flow is a canceller of current subscription

Subscribe to all channels from list. Respects redis pattern syntax. Note: Subscription won't be unsubscribed automatically on thread end. Use canceller explicitly to cancel subscription

withModifiedRuntime

Arguments

:: HasCallStack

=> (FlowRuntime -> FlowRuntime)
Temporary modification function for runtime

-> Flow a
Computation to run with modified runtime

-> m a

Run a flow with a modified runtime. The runtime will be restored after the computation finishes.

Since: 2.0.3.1

type ReaderFlow r = ReaderT r Flow

logging

logCallStack :: (HasCallStack, MonadFlow m) => m ()

logExceptionCallStack :: (HasCallStack, Exception e, MonadFlow m) => e -> m ()

logInfo :: forall (tag :: Type) (m :: Type -> Type). (HasCallStack, MonadFlow m, Show tag, Typeable tag) => tag -> Text -> m ()

logError :: forall (tag :: Type) (m :: Type -> Type). (HasCallStack, MonadFlow m, Show tag, Typeable tag) => tag -> Text -> m ()

logDebug :: forall (tag :: Type) (m :: Type -> Type). (HasCallStack, MonadFlow m, Show tag, Typeable tag) => tag -> Text -> m ()

logWarning :: forall (tag :: Type) (m :: Type -> Type). (HasCallStack, MonadFlow m, Show tag, Typeable tag) => tag -> Text -> m ()

logInfoM :: forall (tag :: Type) (m :: Type -> Type) msg val. (HasCallStack, MonadFlow m, Show tag, Typeable tag, ToJSON msg, ToJSON val) => tag -> msg -> val -> m ()

Log message with Info level.

Thread safe.

logErrorM :: forall (tag :: Type) (m :: Type -> Type) msg val. (HasCallStack, MonadFlow m, Show tag, Typeable tag, ToJSON msg, ToJSON val) => tag -> msg -> val -> m ()

Log message with Error level.

Thread safe.

logDebugM :: forall (tag :: Type) (m :: Type -> Type) msg val. (HasCallStack, MonadFlow m, Show tag, Typeable tag, ToJSON msg, ToJSON val) => tag -> msg -> val -> m ()

Log message with Debug level.

Thread safe.

logWarningM :: forall (tag :: Type) (m :: Type -> Type) msg val. (HasCallStack, MonadFlow m, Show tag, Typeable tag, ToJSON msg, ToJSON val) => tag -> msg -> val -> m ()

Log message with Warning level.

Thread safe.

logInfoV :: forall (tag :: Type) (m :: Type -> Type) val. (HasCallStack, MonadFlow m, Show tag, Typeable tag, ToJSON val) => tag -> val -> m ()

logErrorV :: forall (tag :: Type) (m :: Type -> Type) val. (HasCallStack, MonadFlow m, Show tag, Typeable tag, ToJSON val) => tag -> val -> m ()

logDebugV :: forall (tag :: Type) (m :: Type -> Type) val. (HasCallStack, MonadFlow m, Show tag, Typeable tag, ToJSON val) => tag -> val -> m ()

logWarningV :: forall (tag :: Type) (m :: Type -> Type) val. (HasCallStack, MonadFlow m, Show tag, Typeable tag, ToJSON val) => tag -> val -> m ()

Calling external services

callAPI :: (HasCallStack, MonadFlow m) => BaseUrl -> EulerClient a -> m (Either ClientError a)

The same as callAPI` but with default manager to be used.

callAPI' :: (HasCallStack, MonadFlow m) => Maybe ManagerSelector -> BaseUrl -> EulerClient a -> m (Either ClientError a)

Method for calling external HTTP APIs using the facilities of servant-client. Allows to specify what manager should be used. If no manager found, HttpManagerNotFound will be returne (as part of ConnectionError).

Thread safe, exception free.

Alias for callServantAPI.

| Takes remote url, servant client for this endpoint and returns either client error or result.

data User = User { firstName :: String, lastName :: String , userGUID :: String}
  deriving (Generic, Show, Eq, ToJSON, FromJSON )

data Book = Book { author :: String, name :: String }
  deriving (Generic, Show, Eq, ToJSON, FromJSON )

type API = "user" :> Get '[JSON] User
      :<|> "book" :> Get '[JSON] Book

api :: HasCallStack => Proxy API
api = Proxy

getUser :: HasCallStack => EulerClient User
getBook :: HasCallStack => EulerClient Book
(getUser :<|> getBook) = client api

url = BaseUrl Http "localhost" port ""

myFlow = do
  book <- callAPI url getBook
  user <- callAPI url getUser

callHTTP :: (HasCallStack, MonadFlow m) => HTTPRequest -> m (Either Text HTTPResponse)

The same as callHTTP' but uses the default HTTP manager.

Thread safe, exception free.

Takes remote url and returns either client error or result.

myFlow = do
  book <- callHTTP url

callHTTP'

Arguments

:: (HasCallStack, MonadFlow m)

=> Maybe ManagerSelector
Selector

-> HTTPRequest
remote url Text

-> m (Either Text HTTPResponse)
result

Method for calling external HTTP APIs without bothering with types with custom manager.

Thread safe, exception free.

Takes remote url, optional custom manager selector and returns either client error or result.

myFlow = do
  book <- callHTTPWithManager url mSel

callHTTPWithCert :: MonadFlow m => HTTPRequest -> Maybe HTTPCert -> m (Either Text HTTPResponse)

Deprecated: Use getHTTPManager/callHTTPUsingManager instead. This method does not allow custom CA store.

callHTTPWithManager

Arguments

:: (HasCallStack, MonadFlow m)

=> Maybe ManagerSelector
Selector

-> HTTPRequest
remote url Text

-> m (Either Text HTTPResponse)
result

Deprecated: Use callHTTP' instead. This method has a confusing name, as it accepts a selector not a manager.

other

runIO :: (HasCallStack, MonadFlow m) => IO a -> m a

Run some IO operation, result should have ToJSONEx instance (extended ToJSON), because we have to collect it in recordings for ART system.

Warning. This method is dangerous and should be used wisely.

myFlow = do
  content <- runIO $ readFromFile file
  logDebugT "content id" $ extractContentId content
  pure content

withRunFlow :: ((forall x. Flow x -> IO x) -> IO a) -> Flow a

MonadBaseControl/UnliftIO-like interface for flow.

withSomeResourceFromIO :: (SomeRes -> IO a) -> IO a
someFlowAction :: SomeRes -> Flow Result

example :: Flow Result
example = do
  withRunFlow \runFlow -> do
    withSomeResourceFromIO \res -> do
      runFlow (someFlowAction res)

forkFlow :: HasCallStack => Description -> Flow () -> Flow ()

Fork a unit-returning flow.

Note: to fork a flow which yields a value use 'forkFlow'' instead.

Warning: With forked flows, race coniditions and dead / live blocking become possible. All the rules applied to forked threads in Haskell can be applied to forked flows.

Generally, the method is thread safe. Doesn't do anything to bookkeep the threads. There is no possibility to kill a thread at the moment.

Thread safe, exception free.

myFlow1 = do
  logInfoT "myflow1" "logFromMyFlow1"
  someAction

myFlow2 = do
  _ <- runIO someAction
  forkFlow "myFlow1 fork" myFlow1
  pure ()

forkFlow' :: HasCallStack => Description -> Flow a -> Flow (Awaitable (Either Text a))

Same as forkFlow, but takes Flow a and returns an Awaitable which can be used to reap results from the flow being forked.

myFlow1 = do
  logInfoT "myflow1" "logFromMyFlow1"
  pure 10

myFlow2 = do
  awaitable <- forkFlow' "myFlow1 fork" myFlow1
  await Nothing awaitable

foldFlow :: Monad m => (forall b. FlowMethod b -> m b) -> Flow a -> m a