EulerHS.Runtime

Core Runtime

newtype CoreRuntime

Constructors

CoreRuntime

Fields

_loggerRuntime :: LoggerRuntime

data LoggerRuntime

Constructors

LoggerRuntime

Fields

_flowFormatter :: FlowFormatter

_logContext :: LogContext

_logLevel :: LogLevel

_logRawSql :: ShouldLogSQL

_logCounter :: !LogCounter

_logMaskingConfig :: Maybe LogMaskingConfig

_logLoggerHandle :: LoggerHandle

MemoryLoggerRuntime !FlowFormatter LogContext !LogLevel !(MVar [Text]) !LogCounter

shouldLogRawSql :: LoggerRuntime -> Bool

incLogCounter :: LogCounter -> IO Int

createCoreRuntime :: LoggerRuntime -> IO CoreRuntime

createVoidLoggerRuntime :: IO LoggerRuntime

createMemoryLoggerRuntime :: FlowFormatter -> LogLevel -> IO LoggerRuntime

createLoggerRuntime :: FlowFormatter -> LoggerConfig -> IO LoggerRuntime

createLoggerRuntime' :: Maybe DateFormat -> Maybe Renderer -> BufferSize -> FlowFormatter -> LoggerConfig -> IO LoggerRuntime

clearCoreRuntime :: CoreRuntime -> IO ()

clearLoggerRuntime :: LoggerRuntime -> IO ()

getLogMaskingConfig :: LoggerRuntime -> Maybe LogMaskingConfig

withTransaction :: forall beM a. SqlConn beM -> (NativeSqlConn -> IO a) -> IO (Either SomeException a)

Framework Runtime

data FlowRuntime

FlowRuntime state and options.

Constructors

FlowRuntime

Fields

_coreRuntime :: CoreRuntime

Contains logger settings

_defaultHttpClientManager :: Manager

Http default manager, used for external api calls

_httpClientManagers :: HashMap Text Manager

Http managers, used for external api calls

_dynHttpClientManagers :: MVar (LruCache HTTPClientSettings Manager)

LRU cache of Managers.

_options :: MVar (Map Text Any)

Typed key-value storage

_kvdbConnections :: MVar (Map Text NativeKVDBConn)

Connections for key-value databases

_sqldbConnections :: MVar (Map ConnTag NativeSqlPool)

Connections for SQL databases

_pubSubController :: PubSubController

Subscribe controller

_pubSubConnection :: Maybe Connection

Connection being used for Publish

createFlowRuntime :: CoreRuntime -> IO FlowRuntime

Create default FlowRuntime.

createFlowRuntime' :: Maybe (IO LoggerRuntime) -> IO FlowRuntime

withFlowRuntime :: Maybe (IO LoggerRuntime) -> (FlowRuntime -> IO a) -> IO a

Run flow with given logger runtime creation function.

kvDisconnect :: NativeKVDBConn -> IO ()

runPubSubWorker :: FlowRuntime -> (Text -> IO ()) -> IO (IO ())

shouldFlowLogRawSql :: FlowRuntime -> Bool

Registration for self-signed certs

newtype CertificateRegistrationError

Possible issues that can arise when registering certificates.

Since: 2.0.4.3

Constructors

NoCertificatesAtPath FilePath

withSelfSignedFlowRuntime :: HashMap Text FilePath -> Maybe (IO LoggerRuntime) -> (Either CertificateRegistrationError FlowRuntime -> IO a) -> IO a

Deprecated: use manager builders instead, see HttpAPI.hs

Works identically to withFlowRuntime, but takes an extra parameter. This parameter is a map of textual identifiers to paths where custom CA certificates can be found.

You can then use callAPI, providing Just the textual identifier to use the self-signed certificate(s) provided.

The handler is provided an Either to allow for graceful recovery; if you have nothing you can do with a CertificateRegistrationError, you can throw it.

Since: 2.0.4.3