EulerHS.Types

Documentation

class (Typeable k, ToJSON k) => OptionEntity k v | k -> v

mkOptionKey :: forall k v. OptionEntity k v => k -> Text

data RedisConfig

Constructors

RedisConfig

Fields

connectHost :: String

connectPort :: Word16

connectAuth :: Maybe Text

connectDatabase :: Integer

connectMaxConnections :: Int

connectMaxIdleTime :: NominalDiffTime

connectTimeout :: Maybe NominalDiffTime

data KVDBConfig

Constructors

KVDBConfig Text RedisConfig

KVDBClusterConfig Text RedisConfig

newtype NativeKVDBConn

Constructors

NativeKVDB Connection

type KVDBAnswer = Either KVDBReply

data TxResult a

Constructors

TxSuccess a

TxAborted

TxError String

type KVDBStatus = KVDBStatusF ByteString

data KVDBStatusF bs

Constructors

Ok

Pong

Status bs

type KVDBReply = KVDBReplyF ByteString

data KVDBReplyF bs

Constructors

SingleLine bs

Err bs

Integer Integer

Bulk (Maybe bs)

MultiBulk (Maybe [KVDBReplyF bs])

ExceptionMessage String

KVDBError KVDBError String

data KVDBError

Constructors

KVDBConnectionAlreadyExists

KVDBConnectionDoesNotExist

KVDBConnectionFailed

data KVDBConn

Constructors

Redis !Text !Connection

type KVDBKey = Text

fromRdStatus :: Status -> KVDBStatus

fromRdTxResult :: TxResult a -> TxResult a

hedisReplyToKVDBReply :: Reply -> KVDBReply

exceptionToKVDBReply :: Exception e => e -> KVDBReply

kvdbToNative :: KVDBConn -> NativeKVDBConn

Transform KVDBConn to NativeKVDBConn

nativeToKVDB :: Text -> NativeKVDBConn -> KVDBConn

Transforms NativeKVDBConn to KVDBConn

defaultKVDBConnConfig :: RedisConfig

mkKVDBConfig :: Text -> RedisConfig -> KVDBConfig

Create configuration KVDBConfig for Redis

mkKVDBClusterConfig :: Text -> RedisConfig -> KVDBConfig

Create cluster configuration KVDBConfig for Redis

mkRedisConn :: KVDBConfig -> IO KVDBConn

Create KVDBConn from KVDBConfig

data Microseconds

Constructors

Microseconds Word32

data Awaitable s

Constructors

Awaitable (MVar s)

type Description = Text

newtype ManagerSelector

Constructors

ManagerSelector Text

type SafeFlowGUID = Text

type ForkGUID = Text

type FlowGUID = Text

type Log = [LogEntry]

data LogEntry

Constructors

LogEntry !LogLevel !Message

data PendingMsg

Constructors

PendingMsg !(Maybe FlowGUID) !LogLevel !Tag !Message !MessageNumber !LogContext

data LoggerConfig

Constructors

LoggerConfig

Fields

_isAsync :: Bool

_logLevel :: LogLevel

_logFilePath :: FilePath

_logToConsole :: Bool

_logToFile :: Bool

_maxQueueSize :: Word

_logRawSql :: ShouldLogSQL

_logMaskingConfig :: Maybe LogMaskingConfig

type LogContext = HashMap Text Text

type FlowFormatter = Maybe FlowGUID -> IO MessageFormatter

type MessageFormatter = PendingMsg -> MessageBuilder

type BufferSize = Int

type Tag = Text

data Message

Constructors

Message

Fields

msgMessage :: Maybe Value

msgValue :: Maybe Value

type LogCounter = IORef Int

data ShouldLogSQL

Constructors

UnsafeLogSQL_DO_NOT_USE_IN_PRODUCTION
Log SQL queries, including sensitive data and API keys. Do NOT PR code with this enabled, and make sure this doesn't make it into production

SafelyOmitSqlLogs
omit SQL logs

data MaskKeyType

Constructors

WhiteListKey

BlackListKey

data MessageBuilder

Constructors

SimpleString String

SimpleText Text

SimpleBS ByteString

SimpleLBS ByteString

MsgBuilder Builder

MsgTransformer (Msg -> Msg)

data LogMaskingConfig

Constructors

LogMaskingConfig

Fields

_maskKeys :: HashSet Text

_maskText :: Maybe Text

_keyType :: MaskKeyType

data LogLevel

Logging level.

Constructors

Debug

Info

Warning

Error

defaultMessageFormatter :: MessageFormatter

showingMessageFormatter :: MessageFormatter

defaultLoggerConfig :: LoggerConfig

defaultFlowFormatter :: FlowFormatter

builderToByteString :: Builder -> ByteString

shouldMaskKey :: Maybe LogMaskingConfig -> Text -> Bool

defaultMaskText :: Text

maskHTTPHeaders :: (Text -> Bool) -> Text -> Map Text Text -> Map Text Text

maskServantHeaders :: (Text -> Bool) -> Text -> Seq Header -> Seq Header

maskQueryStrings :: (Text -> Bool) -> Text -> Seq QueryItem -> Seq QueryItem

parseRequestResponseBody :: (Text -> Bool) -> Text -> Maybe ByteString -> ByteString -> Value

maskJSON :: (Text -> Bool) -> Text -> Value -> Value

handleObject :: (Text -> Bool) -> Text -> Object -> Object

handleQueryString :: ByteString -> Value

notSupportedPlaceHolder :: Maybe ByteString -> Value

isContentTypeBlockedForLogging :: Maybe ByteString -> Bool

getContentTypeForServant :: ResponseHeaders -> Maybe ByteString

getContentTypeForHTTP :: Map Text Text -> Maybe ByteString

newtype LBinaryString

Constructors

LBinaryString

Fields

getLBinaryString :: ByteString

newtype BinaryString

Constructors

BinaryString

Fields

getBinaryString :: ByteString

base64Encode :: ByteString -> Text

Base64 encode a bytestring

NOTE: Decoding to UTF-8 cannot fail so is safe

base64Decode :: MonadFail m => Text -> m ByteString

Base64 decode a Base64-encoded string

NOTE: This may fail if the string is malformed using MonadFail

data HTTPIOException

Used when some IO (or other) exception ocurred during a request

Constructors

HTTPIOException Text HTTPRequestMasked

data HTTPMethod

Constructors

Get

Put

Post

Delete

Head

Trace

Connect

Options

Patch

data HTTPCert

Constructors

HTTPCert

Fields

getCert :: ByteString

getCertChain :: [ByteString]

getCertHost :: String

Defines the name of the server, along with an extra ^ service identification blob (not supported in Euler ATM). ^ This is important that the hostname part is properly ^ filled for security reason, as it allows to properly ^ as it allow to properly associate the remote side ^ with the given certificate during a handshake.

getCertKey :: ByteString

data HTTPResponse

Constructors

HTTPResponse

Fields

getResponseBody :: LBinaryString

getResponseCode :: Int

getResponseHeaders :: Map HeaderName HeaderValue

getResponseStatus :: Text

data HTTPRequestMasked

data HTTPRequest

Constructors

HTTPRequest

Fields

getRequestMethod :: HTTPMethod

getRequestHeaders :: Map HeaderName HeaderValue

getRequestBody :: Maybe LBinaryString

getRequestURL :: Text

getRequestTimeout :: Maybe Int

timeout, in microseconds

getRequestRedirects :: Maybe Int

data HTTPClientSettings

buildSettings :: HTTPClientSettings -> ManagerSettings

The simplest settings builder

withProxy :: SimpleProxySettings -> HTTPClientSettings

Add unconditional proxying (for both http/https, regardless HTTP.Client's request proxy settings).

withMbProxy :: Maybe SimpleProxySettings -> HTTPClientSettings

The same as withProxy but to use with optionally existsting settings.

withClientTls :: HTTPCert -> HTTPClientSettings

Adds a client certificate to do client's TLS authentication

withMbClientTls :: Maybe HTTPCert -> HTTPClientSettings

withCustomCA :: CertificateStore -> HTTPClientSettings

Adds an additional store with trusted CA certificates. There is no Maybe version since CertificateStore is a monoid.

makeCertificateStoreFromMemory :: [ByteString] -> Either String CertificateStore

Make a store from in-memory certs

withNoCheckLeafV3 :: HTTPClientSettings

Warning: Don't use in production code, use X509 v3 certs instead.

Turns off the check that all certs are X509 v3 ones

mkHttpApiCallLogEntry :: Integer -> HTTPRequestMasked -> HTTPResponseMasked -> HttpApiCallLogEntry

httpGet :: Text -> HTTPRequest

HTTP GET request.

httpGet "https://google.com"

httpPut :: Text -> HTTPRequest

httpPost :: Text -> HTTPRequest

httpDelete :: Text -> HTTPRequest

httpHead :: Text -> HTTPRequest

defaultRequest :: HTTPMethod -> Text -> HTTPRequest

defaultTimeout :: Int

withHeader :: HeaderName -> HeaderValue -> HTTPRequest -> HTTPRequest

Add a header to an HTTPRequest

httpGet "https://google.com"
  & withHeader "Content-Type" "application/json"

withOptionalHeader :: HeaderName -> Maybe HeaderValue -> HTTPRequest -> HTTPRequest

withTimeout :: Int -> HTTPRequest -> HTTPRequest

Sets timeout, in microseconds

withRedirects :: Int -> HTTPRequest -> HTTPRequest

withBody :: [(Text, Text)] -> HTTPRequest -> HTTPRequest

extractBody :: HTTPResponse -> Text

maskHTTPRequest :: Maybe LogMaskingConfig -> HTTPRequest -> HTTPRequestMasked

maskHTTPResponse :: Maybe LogMaskingConfig -> HTTPResponse -> HTTPResponseMasked

data ServantApiCallLogEntry

Constructors

ServantApiCallLogEntry

Fields

url :: String

method :: Text

req_headers :: Value

req_body :: Value

res_code :: Int

res_body :: Value

res_headers :: Value

latency :: Integer

data LogServantResponse

Constructors

LogServantResponse

Fields

statusCode :: String

headers :: Seq (Text, Text)

httpVersion :: String

body :: Value

data LogServantRequest

Constructors

LogServantRequest

Fields

url :: String

method :: Text

body :: Value

headers :: Seq (Text, Text)

queryString :: Seq (Text, Maybe Text)

newtype EulerClient a

Constructors

EulerClient (Free ClientF a)

mkServantApiCallLogEntry :: Maybe LogMaskingConfig -> BaseUrl -> Request -> Response -> Integer -> ServantApiCallLogEntry

client :: HasClient EulerClient api => Proxy api -> Client EulerClient api

interpretClientF :: (forall msg. ToJSON msg => msg -> IO ()) -> Maybe LogMaskingConfig -> BaseUrl -> ClientF a -> ClientM a

runEulerClient :: (forall msg. ToJSON msg => msg -> IO ()) -> Maybe LogMaskingConfig -> BaseUrl -> EulerClient a -> ClientM a

data MySQLConfig

Since: 2.0.3.0

Constructors

MySQLConfig

Fields

connectHost :: String

connectPort :: Word16

connectUser :: String

connectPassword :: String

connectDatabase :: String

connectOptions :: [MySqlOption]

connectPath :: FilePath

connectSSL :: Maybe SSLInfo

connectCharset :: !MySQLCharset

Since: 2.0.3.0

data MySQLCharset

Describes the character set to be used with the database. This also includes collation information.

Currently, only a limited number of these are provided.

See also: MySQL documentation on character sets

Since: 2.0.3.0

Constructors

Latin1
Corresponds to the latin1 character set, with the latin1_swedish_ci collation.

Since: 2.0.3.0

UTF8General
Corresponds to the utf8 character set, with the utf8_general_ci collation.

Since: 2.0.3.0

UTF8Full
Corresponds to the utf8mb character set, with the unicode_ci collation.

Since: 2.0.3.0

data MySqlOption

Constructors

ConnectTimeout Word

Compress

NamedPipe

ReadDefaultFile FilePath
InitCommand ByteString -- TODO

CharsetDir FilePath
ReadDefaultGroup ByteString -- TODO

CharsetName String

LocalInFile Bool

Protocol MySqlProtocol

ReadTimeout Word
SharedMemoryBaseName ByteString -- TODO

WriteTimeout Word

SecureAuth Bool
UseRemoteConnection | UseEmbeddedConnection | GuessConnection | ClientIP ByteString

ReportDataTruncation Bool

Reconnect Bool

FoundRows
SSLVerifyServerCert Bool

IgnoreSIGPIPE

IgnoreSpace

Interactive

LocalFiles

MultiResults

MultiStatements

NoSchema

defaultMySQLConfig :: MySQLConfig

Since: 2.0.3.0

createMySQLConn :: MySQLConfig -> IO MySQLConn

Connect with the given config to the database.

Since: 2.0.3.0

closeMySQLConn :: MySQLConn -> IO ()

Close the given connection.

data PostgresConfig

Constructors

PostgresConfig

Fields

connectHost :: String

connectPort :: Word16

connectUser :: String

connectPassword :: String

connectDatabase :: String

createPostgresConn :: PostgresConfig -> IO Connection

Connect with the given config to the database.

closePostgresConn :: Connection -> IO ()

Close the given connection.

type DBResult a = Either DBError a

Represents resulting type for DB actions

data DBError

Represents DB error

Constructors

DBError DBErrorType Text

data DBErrorType

Represents failures that may occur while working with the database

Constructors

ConnectionFailed

ConnectionAlreadyExists

ConnectionDoesNotExist

TransactionRollbacked

SQLError SQLError

UnexpectedResult

UnrecognizedError

data PostgresSqlError

Constructors

PostgresSqlError

Fields

sqlState :: Text

sqlExecStatus :: PostgresExecStatus

sqlErrorMsg :: Text

sqlErrorDetail :: Text

sqlErrorHint :: Text

data PostgresExecStatus

Constructors

PostgresEmptyQuery

PostgresCommandOk

PostgresTuplesOk

PostgresCopyOut

PostgresCopyIn

PostgresCopyBoth

PostgresBadResponse

PostgresNonfatalError

PostgresFatalError

PostgresSingleTuple

data MysqlSqlError

Constructors

MysqlSqlError

Fields

errCode :: !Word16

errMsg :: !Text

data SQLError

Constructors

PostgresError PostgresSqlError

MysqlError MysqlSqlError

SqliteError SqliteSqlError

data SqliteSqlError

Constructors

SqliteSqlError

Fields

sqlError :: !SqliteError

sqlErrorDetails :: Text

sqlErrorContext :: Text

data SqliteError

Constructors

SqliteErrorOK

SqliteErrorError

SqliteErrorInternal

SqliteErrorPermission

SqliteErrorAbort

SqliteErrorBusy

SqliteErrorLocked

SqliteErrorNoMemory

SqliteErrorReadOnly

SqliteErrorInterrupt

SqliteErrorIO

SqliteErrorCorrupt

SqliteErrorNotFound

SqliteErrorFull

SqliteErrorCantOpen

SqliteErrorProtocol

SqliteErrorEmpty

SqliteErrorSchema

SqliteErrorTooBig

SqliteErrorConstraint

SqliteErrorMismatch

SqliteErrorMisuse

SqliteErrorNoLargeFileSupport

SqliteErrorAuthorization

SqliteErrorFormat

SqliteErrorRange

SqliteErrorNotADatabase

SqliteErrorNotice

SqliteErrorWarning

SqliteErrorRow

SqliteErrorDone

data PoolConfig

Represents Pool parameters

Constructors

PoolConfig

Fields

stripes :: Int

a number of sub-pools

keepAlive :: NominalDiffTime

the amount of time the connection will be stored

resourcesPerStripe :: Int

maximum number of connections to be stored in each sub-pool

data DBConfig (beM :: Type -> Type)

Represents DB configurations

Constructors

PostgresPoolConf ConnTag PostgresConfig PoolConfig
config for Pool with Postgres connections

MySQLPoolConf ConnTag MySQLConfig PoolConfig
config for Pool with MySQL connections

SQLitePoolConf ConnTag SQliteDBname PoolConfig
config for Pool with SQlite connections

data SqlConn (beM :: Type -> Type)

Represents SQL connection that we use in flow. Parametrised by BEAM monad corresponding to the certain DB (MySQL, Postgres, SQLite)

Constructors

PostgresPool ConnTag (Pool Connection)
Pool with Postgres connections

MySQLPool ConnTag (Pool MySQLConn)
Pool with MySQL connections

SQLitePool ConnTag (Pool Connection)
Pool with SQLite connections

type SQliteDBname = String

Represents path to the SQLite DB

type ConnTag = Text

Tag for SQL connections

data NativeSqlConn

Representation of native DB connections that we use in implementation.

Constructors

NativePGConn Connection

NativeMySQLConn MySQLConn

NativeSQLiteConn Connection

data NativeSqlPool

Representation of native DB pools that we store in FlowRuntime

Constructors

NativePGPool (Pool Connection)
Pool with Postgres connections

NativeMySQLPool (Pool MySQLConn)
Pool with MySQL connections

NativeSQLitePool (Pool Connection)
Pool with SQLite connections

class BeamRunner beM where

Methods

getBeamDebugRunner :: NativeSqlConn -> beM a -> (Text -> IO ()) -> IO a

class (BeamSqlBackend be, MonadBeam be beM) => BeamRuntime be beM | be -> beM, beM -> be where

Methods

rtSelectReturningList :: FromBackendRow be a => SqlSelect be a -> beM [a]

rtSelectReturningOne :: FromBackendRow be a => SqlSelect be a -> beM (Maybe a)

rtInsert :: SqlInsert be table -> beM ()

rtInsertReturningList :: forall table. (Beamable table, FromBackendRow be (table Identity)) => SqlInsert be table -> beM [table Identity]

rtUpdate :: SqlUpdate be table -> beM ()

rtUpdateReturningList :: forall table. (Beamable table, FromBackendRow be (table Identity)) => SqlUpdate be table -> beM [table Identity]

rtDelete :: SqlDelete be table -> beM ()

deleteReturningListPG :: (Beamable table, FromBackendRow Postgres (table Identity)) => SqlDelete Postgres table -> Pg [table Identity]

updateReturningListPG :: (Beamable table, FromBackendRow Postgres (table Identity)) => SqlUpdate Postgres table -> Pg [table Identity]

bemToNative :: SqlConn beM -> NativeSqlPool

Transform SqlConn to NativeSqlPool

mkSqlConn :: DBConfig beM -> IO (SqlConn beM)

Create SqlConn from DBConfig

mkSQLiteConfig :: ConnTag -> SQliteDBname -> DBConfig SqliteM

Create SQLite DBConfig

mkSQLitePoolConfig :: ConnTag -> SQliteDBname -> PoolConfig -> DBConfig SqliteM

Create SQLite Pool DBConfig

mkPostgresConfig :: ConnTag -> PostgresConfig -> DBConfig Pg

Create Postgres DBConfig

mkPostgresPoolConfig :: ConnTag -> PostgresConfig -> PoolConfig -> DBConfig Pg

Create Postgres Pool DBConfig

mkMySQLConfig :: ConnTag -> MySQLConfig -> DBConfig MySQLM

Create MySQL DBConfig

mkMySQLPoolConfig :: ConnTag -> MySQLConfig -> PoolConfig -> DBConfig MySQLM

Create MySQL Pool DBConfig

getDBName :: DBConfig beM -> String

sqliteErrorToDbError :: Text -> SQLError -> DBError

mysqlErrorToDbError :: Text -> ERRException -> DBError

postgresErrorToDbError :: Text -> SqlError -> DBError

nativeToBem :: ConnTag -> NativeSqlPool -> SqlConn beM

Transforms NativeSqlPool to SqlConn

newtype HttpManagerNotFound

Constructors

HttpManagerNotFound Text

data AwaitingError

Constructors

AwaitingTimeout

ForkedFlowError Text